+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add support for links of all item kinds (enum variant, macro attribute, etc.)
+- Don't print extraneous newline
 
 ## [0.2.1] - 2025-07-10