+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- next-header -->
 ## [Unreleased]
 
+## [0.3.0] - 2025-07-11
+
 ### Added
 
 - Add support for links of all item kinds (enum variant, macro attribute, etc.)
@@ -50,7 +52,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.0] - 2025-07-09
 
 <!-- next-url -->
-[Unreleased]: https://github.com/bluurryy/cargo-insert-docs/compare/v0.2.1...HEAD
+[Unreleased]: https://github.com/bluurryy/cargo-insert-docs/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/bluurryy/cargo-insert-docs/compare/v0.2.1...v0.3.0
 [0.2.1]: https://github.com/bluurryy/cargo-insert-docs/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/bluurryy/cargo-insert-docs/compare/v0.1.2...v0.2.0
 [0.1.2]: https://github.com/bluurryy/cargo-insert-docs/compare/v0.1.1...v0.1.2